Split inline snapshots to their own file
