+++ testing/snapshot.ts
@@ -71,30 +71,6 @@
  * deno test --allow-all -- --update
  * ```
  *
- * For inline snapshots, using an `expectedSnapshot` of the template literal
- * \`CREATE\` will mark a snapshot for creation. This template literal must not
- * appear elsewhere in the file, as the updater uses it to determine where to
- * place new snapshots.
- *
- * ```ts
- * // example_test.ts
- * import { assertInlineSnapshot } from "@std/testing/snapshot";
- *
- * Deno.test("isInlineSnapshotMatch", async function (t): Promise<void> {
- *   const a = {
- *     hello: "world!",
- *     example: 123,
- *   };
- *   await assertInlineSnapshot(t, a, `UPDATE`);
- * });
- * ```
- *
- * Inline snapshots do not use the update flag.
- *
- * ```sh
- * deno test --allow-all
- * ```
- *
  * ## Permissions:
  *
  * When running `assertSnapshot`, the `--allow-read` permission must be enabled, or
@@ -200,9 +176,14 @@ import { ensureFile, ensureFileSync } from "@std/fs/ensure-file";
 import { assert } from "@std/assert/assert";
 import { AssertionError } from "@std/assert/assertion-error";
 import { equal } from "@std/assert/equal";
-import { diff } from "@std/internal/diff";
-import { diffStr } from "@std/internal/diff-str";
-import { buildMessage } from "@std/internal/build-message";
+import {
+  escapeStringForJs,
+  getErrorMessage,
+  getIsUpdate,
+  getOptions,
+  getSnapshotNotMatchMessage,
+  serialize,
+} from "./_snapshot_utils.ts";
 
 const SNAPSHOT_DIR = "__snapshots__";
 const SNAPSHOT_EXT = "snap";
@@ -248,115 +229,6 @@ export interface SnapshotOptions<T = unknown> {
    * Function to use when serializing the snapshot. The default is {@linkcode serialize}.
    */
   serializer?: (actual: T) => string;
-}
-
-/** The options for {@linkcode assertInlineSnapshot}. */
-export interface InlineSnapshotOptions<T = unknown>
-  extends Pick<SnapshotOptions<T>, "msg" | "serializer"> {
-  /**
-   * Whether to format the test file after updating.
-   *
-   * The default is `true`. If multiple snapshots will be created in one test file
-   * and the tests have incompatible `format` options, the snapshots will be written,
-   * the file will not be formatted, and we will throw.
-   */
-  format?: boolean;
-}
-
-function getErrorMessage(message: string, options: SnapshotOptions) {
-  return typeof options.msg === "string" ? options.msg : message;
-}
-
-/**
- * Default serializer for `assertSnapshot`.
- *
- * @example Usage
- * ```ts
- * import { serialize } from "@std/testing/snapshot";
- * import { assertEquals } from "@std/assert";
- *
- * assertEquals(serialize({ foo: 42 }), "{\n  foo: 42,\n}")
- * ```
- *
- * @param actual The value to serialize
- * @returns The serialized string
- */
-export function serialize(actual: unknown): string {
-  return Deno.inspect(actual, {
-    depth: Infinity,
-    sorted: true,
-    trailingComma: true,
-    compact: false,
-    iterableLimit: Infinity,
-    strAbbreviateSize: Infinity,
-    breakLength: Infinity,
-    escapeSequences: false,
-  }).replaceAll("\r", "\\r");
-}
-
-/**
- * Converts a string to a valid JavaScript string which can be wrapped in backticks.
- *
- * @example
- *
- * "special characters (\ ` $) will be escaped" -> "special characters (\\ \` \$) will be escaped"
- */
-function escapeStringForJs(str: string) {
-  return str
-    .replace(/\\/g, "\\\\")
-    .replace(/`/g, "\\`")
-    .replace(/\$/g, "\\$");
-}
-
-let _mode: SnapshotMode;
-/**
- * Get the snapshot mode.
- */
-function getMode(options: SnapshotOptions) {
-  if (options.mode) {
-    return options.mode;
-  } else if (_mode) {
-    return _mode;
-  } else {
-    _mode = Deno.args.some((arg) => arg === "--update" || arg === "-u")
-      ? "update"
-      : "assert";
-    return _mode;
-  }
-}
-
-/**
- * Return `true` when snapshot mode is `update`.
- */
-function getIsUpdate(options: SnapshotOptions) {
-  return getMode(options) === "update";
-}
-
-function getOptions<T>(
-  msgOrOpts?: string | T,
-): T {
-  if (msgOrOpts === undefined) return {} as T;
-
-  if (typeof msgOrOpts === "object" && msgOrOpts !== null) {
-    return msgOrOpts;
-  }
-
-  return { msg: msgOrOpts } as T;
-}
-
-function getSnapshotNotMatchMessage(
-  actualSnapshot: string,
-  expectedSnapshot: string,
-  options: SnapshotOptions,
-) {
-  const stringDiff = !actualSnapshot.includes("\n");
-  const diffResult = stringDiff
-    ? diffStr(actualSnapshot, expectedSnapshot)
-    : diff(actualSnapshot.split("\n"), expectedSnapshot.split("\n"));
-  const diffMsg = buildMessage(diffResult, { stringDiff }).join("\n");
-  const message =
-    `Snapshot does not match:\n${diffMsg}\nTo update snapshots, run\n    deno test --allow-read --allow-write [files]... -- --update\n`;
-  return getErrorMessage(message, options);
 }
 
 class AssertSnapshotContext {
@@ -619,159 +491,6 @@ class AssertSnapshotContext {
   }
 }
 
-class AssertInlineSnapshotContext {
-  static contexts = new Map<string, AssertInlineSnapshotContext>();
-
-  /**
-   * Returns an instance of `AssertInlineSnapshotContext`. This will be retrieved from
-   * a cache if an instance was already created for a given test file path.
-   */
-  static fromContext(
-    testContext: Deno.TestContext,
-  ): AssertInlineSnapshotContext {
-    const testFilePath = fromFileUrl(testContext.origin);
-    const { dir, base } = parse(testFilePath);
-    const path = resolve(dir, base);
-
-    let context = this.contexts.get(path);
-    if (context) {
-      return context;
-    }
-
-    context = new this(toFileUrl(path));
-    this.contexts.set(path, context);
-    return context;
-  }
-
-  #teardownRegistered = false;
-  #indexToSnapshot: string[] = [];
-  #snapshotsCreated = 0;
-  #testFileUrl: URL;
-  #format: boolean | undefined | "error" = undefined;
-
-  constructor(testFileUrl: URL) {
-    this.#testFileUrl = testFileUrl;
-  }
-
-  /**
-   * Write updates to the snapshot file and log statistics.
-   */
-  #teardown = () => {
-    if (this.#snapshotsCreated === 0) return;
-    const currentSnapshots = this.#indexToSnapshot;
-
-    const testFilePath = fromFileUrl(this.#testFileUrl);
-    ensureFileSync(testFilePath);
-    const file = Deno.readTextFileSync(testFilePath);
-
-    const parts = file.split("`CREATE`");
-    if (parts.length !== this.#snapshotsCreated + 1) {
-      throw new Error(
-        `assertInlineSnapshot expected to update ${this.#snapshotsCreated} ${
-          this.#snapshotsCreated === 1 ? "snapshot" : "snapshots"
-        } but found ${parts.length - 1} snapshot ${
-          parts.length === 2 ? "location" : "locations"
-        }.`,
-      );
-    }
-
-    let result = parts[0]!;
-
-    for (let i = 0; i < this.#snapshotsCreated; i++) {
-      const createdSnapshot = currentSnapshots[i];
-      if (createdSnapshot === undefined) {
-        throw new Error(
-          `assertInlineSnapshot expected to create a snapshot at index ${i} but none was registered.`,
-        );
-      }
-
-      const formattedSnapshot = escapeStringForJs(createdSnapshot);
-      result += `\`${formattedSnapshot}\`${parts[i + 1]}`;
-    }
-
-    Deno.writeTextFileSync(testFilePath, result);
-
-    if (this.#format === undefined || this.#format === true) {
-      const command = new Deno.Command(Deno.execPath(), {
-        args: ["fmt", testFilePath],
-      });
-      const { stderr, success } = command.outputSync();
-      if (!success) {
-        throw new Error(
-          `assertInlineSnapshot errored while formatting ${testFilePath}:\n${
-            new TextDecoder().decode(stderr)
-          }`,
-        );
-      }
-    } else if (this.#format === "error") {
-      throw new Error(
-        "assertInlineSnapshot was called with incompatible format options. Snapshots were added but the file was not formatted.",
-      );
-    }
-
-    const created = this.#snapshotsCreated;
-    if (created > 0) {
-      // deno-lint-ignore no-console
-      console.log(
-        `%c\n > ${created} ${
-          created === 1 ? "snapshot" : "snapshots"
-        } created.`,
-        "color: green; font-weight: bold;",
-      );
-    }
-  };
-
-  /**
-   * Register a teardown function which writes the snapshot file to disk and logs the number
-   * of snapshots updated after all tests have run.
-   *
-   * This method can safely be called more than once and will only register the teardown
-   * function once in a context.
-   */
-  async registerTeardown() {
-    if (!this.#teardownRegistered) {
-      for (const perm of ["read", "write"] as const) {
-        const permission = await Deno.permissions.query({
-          name: perm,
-          path: this.#testFileUrl,
-        });
-        if (permission.state !== "granted") {
-          throw new Deno.errors.PermissionDenied(
-            `Missing ${perm} access to snapshot file (${this.#testFileUrl}). This is required because assertInlineSnapshot is trying to create snapshots. Please pass the --allow-${perm} flag.`,
-          );
-        }
-      }
-      globalThis.addEventListener("unload", this.#teardown);
-      this.#teardownRegistered = true;
-    }
-  }
-
-  /**
-   * Gets the number of snapshots which have been created and increments the count by 1.
-   */
-  getCount() {
-    const count = this.#snapshotsCreated;
-    this.#snapshotsCreated++;
-    return count;
-  }
-
-  /**
-   * Creates a snapshot by index. Updates will be written to the test file when all
-   * tests have run.
-   */
-  createSnapshot(index: number, snapshot: string, format: boolean | undefined) {
-    this.#indexToSnapshot[index] = snapshot;
-
-    if (format === undefined) format = true;
-
-    if (this.#format === undefined) {
-      this.#format = format;
-    } else if (this.#format !== format) {
-      this.#format = "error";
-    }
-  }
-}
-
 /**
  * Make an assertion that `actual` matches a snapshot. If the snapshot and `actual` do
  * not match, then throw.
@@ -925,138 +644,5 @@ export function createAssertSnapshot<T>(
     };
 
     await baseAssertSnapshot(context, actual, mergedOptions);
-  };
-}
-
-/**
- * Make an assertion that `actual` matches `expectedSnapshot`. If they do not match,
- * then throw.
- *
- * Type parameter can be specified to ensure values under comparison have the same type.
- *
- * @example Usage
- * ```ts
- * import { assertInlineSnapshot } from "@std/testing/snapshot";
- *
- * Deno.test("snapshot", async (t) => {
- *   await assertInlineSnapshot<number>(t, 2, `2`);
- * });
- * ```
- * @typeParam T The type of the snapshot
- * @param context The test context
- * @param actual The actual value to compare
- * @param expectedSnapshot The expected snapshot, or \`CREATE\` to create
- * @param options The options
- */
-export async function assertInlineSnapshot<T>(
-  context: Deno.TestContext,
-  actual: T,
-  expectedSnapshot: string,
-  options?: InlineSnapshotOptions<T>,
-): Promise<void>;
-/**
- * Make an assertion that `actual` matches `expectedSnapshot`. If they do not match,
- * then throw.
- *
- * Type parameter can be specified to ensure values under comparison have the same type.
- *
- * @example Usage
- * ```ts
- * import { assertInlineSnapshot } from "@std/testing/snapshot";
- *
- * Deno.test("snapshot", async (t) => {
- *   await assertInlineSnapshot<number>(t, 2, `2`);
- * });
- * ```
- * @typeParam T The type of the snapshot
- * @param context The test context
- * @param actual The actual value to compare
- * @param expectedSnapshot The expected snapshot, or \`CREATE\` to create
- * @param message The optional assertion message
- */
-export async function assertInlineSnapshot<T>(
-  context: Deno.TestContext,
-  actual: T,
-  expectedSnapshot: string,
-  message?: string,
-): Promise<void>;
-export async function assertInlineSnapshot(
-  context: Deno.TestContext,
-  actual: unknown,
-  expectedSnapshot: string,
-  msgOrOpts?: string | InlineSnapshotOptions<unknown>,
-) {
-  const options = getOptions(msgOrOpts);
-
-  const serializer = options.serializer ?? serialize;
-  const actualSnapshot = serializer(actual);
-  // TODO(WWRS): dedent expectedSnapshot to allow snapshots to look nicer
-
-  if (expectedSnapshot === `CREATE`) {
-    const assertInlineSnapshotContext = AssertInlineSnapshotContext.fromContext(
-      context,
-    );
-    const index = assertInlineSnapshotContext.getCount();
-    await assertInlineSnapshotContext.registerTeardown();
-    assertInlineSnapshotContext.createSnapshot(
-      index,
-      actualSnapshot,
-      options.format,
-    );
-  } else if (!equal(actualSnapshot, expectedSnapshot)) {
-    throw new AssertionError(
-      getSnapshotNotMatchMessage(actualSnapshot, expectedSnapshot, options),
-    );
-  }
-}
-
-/**
- * Create {@linkcode assertInlineSnapshot} function with the given options.
- *
- * The specified option becomes the default for returned {@linkcode assertInlineSnapshot}
- *
- * @example Usage
- * ```ts
- * import { createAssertInlineSnapshot } from "@std/testing/snapshot";
- *
- * const assertInlineSnapshot = createAssertInlineSnapshot({
- *   // Never format the test file after writing new snapshots
- *   format: false
- * });
- *
- * Deno.test("a snapshot test case", async (t) => {
- *   await assertInlineSnapshot(
- *     t,
- *     { foo: "Hello", bar: "World" },
- *     `CREATE`
- *   );
- * })
- * ```
- *
- * @typeParam T The type of the snapshot
- * @param options The options
- * @param baseAssertSnapshot {@linkcode assertInlineSnapshot} function implementation. Default to the original {@linkcode assertInlineSnapshot}
- * @returns {@linkcode assertInlineSnapshot} function with the given default options.
- */
-export function createAssertInlineSnapshot<T>(
-  options: InlineSnapshotOptions<T>,
-  baseAssertSnapshot: typeof assertInlineSnapshot = assertInlineSnapshot,
-): typeof assertInlineSnapshot {
-  return async function (
-    context: Deno.TestContext,
-    actual: T,
-    expectedSnapshot: string,
-    messageOrOptions?: string | InlineSnapshotOptions<T>,
-  ) {
-    const mergedOptions: InlineSnapshotOptions<T> = {
-      ...options,
-      ...(typeof messageOrOptions === "string"
-        ? {
-          msg: messageOrOptions,
-        }
-        : messageOrOptions),
-    };
-
-    await baseAssertSnapshot(context, actual, expectedSnapshot, mergedOptions);
   };
 }
